requirements.txt
fastapi
stripe
uvicorn
python-dotenv
pydantic
